Split components from mod.rs
